correct offense in snprintf spec
Webserver cache time is printed as a long int, but was stored as
uint32_t.  It's simplest to change the type to long, since it's
programmer-supplied in a struct initializer.  Removing the unsigned
flavor allows for -1, which could signify "forever" more effectively
than the test for zero now in place.  But that's for another commit.

+++ include/webserver.h
@@ -53,7 +53,7 @@ typedef struct
     const void *asset_start;
     const void *asset_end;
     const char *asset_type;
-    uint32_t cache_time; // Cache time in seconds
+    long cache_time; // Cache time in seconds
 } asset_entry_t;
 
 // Lookup table array
